tests/EvmAcceptanceTests/scripts/requirements.txt
cython
cytoolz
eth-brownie
py-solc==3.2.0
py-solc-x==1.1.1
pyzil
jsonrpcclient==3.3.6
web3~=5.11.1
protobuf~=3.20.1
eth-account~=0.5.2
